backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
apscheduler==3.10.4
emails==0.6
jinja2==3.1.4
openpyxl==3.1.5
reportlab==4.2.2
python-dotenv==1.0.1
httpx==0.27.0
slowapi==0.1.9
cryptography==42.0.8
# AI Contract Analysis
anthropic==0.34.2
pdfplumber==0.11.4
